feat: make map landing page, mv current to about
