+++ public/locales/de/menu.json
@@ -4,6 +4,7 @@
   "explore_label": "Erkunden",
   "profile_label": "Profil",
   "settings_label": "Einstellungen",
+  "about_label": "Über OpenSenseMap",
   "my_devices_label": "Meine Geräte",
   "add_device_label": "Gerät hinzufügen",
   "tutorials_label": "Tutorials",
