+++ public/locales/en/menu.json
@@ -3,6 +3,7 @@
   "subtitle": "Please sign in to see more content",
   "explore_label": "Explore",
   "profile_label": "Profile",
+  "about_label": "About",
   "settings_label": "Settings",
   "my_devices_label": "My Devices",
   "add_device_label": "Add Device",
